requirements.txt
langchain==0.0.24
python-dotenv==1.0.0
streamlit
tiktoken==0.4.0
faiss-cpu==1.7.4
langchain_community